+++ packages/teams-js-app-sdk/src/private/chat.ts
@@ -1,7 +1,11 @@
 import { ensureInitialized } from '../internal/internalAPIs';
 import { FrameContexts } from '../public/constants';
 import { OpenConversationRequest } from '../public/interfaces';
-import { sendMessageToParent } from '../internal/communication';
+import {
+  sendAndUnwrap,
+  sendMessageToParent,
+  sendAndHandleStatusAndReason as sendAndHandleError,
+} from '../internal/communication';
 import { registerHandler, removeHandler } from '../internal/handlers';
 import { runtime } from '../public/runtime';
 import { ChatMembersInformation } from './interfaces';
@@ -15,50 +19,44 @@ export namespace chat {
    * Hide from docs
    * --------------
    * Allows the user to start or continue a conversation with each subentity inside the tab
+   * @returns Promise resolved upon completion
    */
-  export function openConversation(openConversationRequest: OpenConversationRequest): void {
-    ensureInitialized(FrameContexts.content);
-    sendMessageToParent(
-      'conversations.openConversation',
-      [
-        {
-          title: openConversationRequest.title,
-          subEntityId: openConversationRequest.subEntityId,
-          conversationId: openConversationRequest.conversationId,
-          channelId: openConversationRequest.channelId,
-          entityId: openConversationRequest.entityId,
-        },
-      ],
-      (status: boolean, reason: string) => {
-        if (!status) {
-          throw new Error(reason);
-        }
-      },
-    );
-    if (openConversationRequest.onStartConversation) {
-      registerHandler(
-        'startConversation',
-        (subEntityId: string, conversationId: string, channelId: string, entityId: string) =>
-          openConversationRequest.onStartConversation({
-            subEntityId,
-            conversationId,
-            channelId,
-            entityId,
-          }),
-      );
-    }
-    if (openConversationRequest.onCloseConversation) {
-      registerHandler(
-        'closeConversation',
-        (subEntityId: string, conversationId?: string, channelId?: string, entityId?: string) =>
-          openConversationRequest.onCloseConversation({
-            subEntityId,
-            conversationId,
-            channelId,
-            entityId,
-          }),
-      );
-    }
+  export function openConversation(openConversationRequest: OpenConversationRequest): Promise<void> {
+    return new Promise<void>(resolve => {
+      ensureInitialized(FrameContexts.content);
+      const sendPromise = sendAndHandleError('conversations.openConversation', {
+        title: openConversationRequest.title,
+        subEntityId: openConversationRequest.subEntityId,
+        conversationId: openConversationRequest.conversationId,
+        channelId: openConversationRequest.channelId,
+        entityId: openConversationRequest.entityId,
+      });
+      if (openConversationRequest.onStartConversation) {
+        registerHandler(
+          'startConversation',
+          (subEntityId: string, conversationId: string, channelId: string, entityId: string) =>
+            openConversationRequest.onStartConversation({
+              subEntityId,
+              conversationId,
+              channelId,
+              entityId,
+            }),
+        );
+      }
+      if (openConversationRequest.onCloseConversation) {
+        registerHandler(
+          'closeConversation',
+          (subEntityId: string, conversationId?: string, channelId?: string, entityId?: string) =>
+            openConversationRequest.onCloseConversation({
+              subEntityId,
+              conversationId,
+              channelId,
+              entityId,
+            }),
+        );
+      }
+      resolve(sendPromise);
+    });
   }
 
   /**
@@ -82,11 +80,13 @@ export namespace chat {
    * Because a malicious party run your content in a browser, this value should
    * be used only as a hint as to who the members are and never as proof of membership.
    * @param callback The callback to invoke when the {@link ChatMembersInformation} object is retrieved.
+   * @returns Promise resolved with information on all chat members
    */
-  export function getChatMembers(callback: (chatMembersInformation: ChatMembersInformation) => void): void {
-    ensureInitialized();
-
-    sendMessageToParent('getChatMembers', callback);
+  export function getChatMembers(): Promise<ChatMembersInformation> {
+    return new Promise<ChatMembersInformation>(resolve => {
+      ensureInitialized();
+      resolve(sendAndUnwrap('getChatMembers'));
+    });
   }
 
   export function isSupported(): boolean {