+++ packages/teams-js-app-sdk/src/public/index.ts
@@ -37,3 +37,4 @@ export { teamsCore } from './teamsAPIs';
 export { people } from './people';
 export { video } from './video';
 export { sharing } from './sharing';
+export { call } from './call';