+++ packages/teams-js-app-sdk/src/public/meeting.ts
@@ -1,7 +1,6 @@
-import { sendMessageToParent } from '../internal/communication';
+import { sendAndHandleSdkError } from '../internal/communication';
 import { registerHandler } from '../internal/handlers';
 import { ensureInitialized } from '../internal/internalAPIs';
-import { SdkError } from './interfaces';
 import { FrameContexts } from './constants';
 import { runtime } from './runtime';
 
@@ -98,6 +97,13 @@ export namespace meeting {
     };
   }
 
+  export interface IAppContentStageSharingCapabilities {
+    /**
+     * indicates whether app has permission to share contents to meeting stage
+     */
+    doesAppHaveSharePermission: boolean;
+  }
+
   export enum MeetingType {
     Unknown = 'Unknown',
     Adhoc = 'Adhoc',
@@ -109,74 +115,58 @@ export namespace meeting {
 
   /**
    * Allows an app to get the incoming audio speaker setting for the meeting user
-   * @param callback Callback contains 2 parameters, error and result.
    * error can either contain an error of type SdkError, incase of an error, or null when fetch is successful
    * result can either contain the true/false value, incase of a successful fetch or null when the fetching fails
-   * result: True means incoming audio is muted and false means incoming audio is unmuted
+   * @returns Promise result where true means incoming audio is muted and false means incoming audio is unmuted
    */
-  export function getIncomingClientAudioState(
-    callback: (error: SdkError | null, result: boolean | null) => void,
-  ): void {
-    if (!callback) {
-      throw new Error('[get incoming client audio state] Callback cannot be null');
-    }
-    ensureInitialized(FrameContexts.sidePanel, FrameContexts.meetingStage);
-    sendMessageToParent('getIncomingClientAudioState', callback);
+  export function getIncomingClientAudioState(): Promise<boolean> {
+    return new Promise<boolean>(resolve => {
+      ensureInitialized(FrameContexts.sidePanel, FrameContexts.meetingStage);
+      resolve(sendAndHandleSdkError('getIncomingClientAudioState'));
+    });
   }
 
   /**
    * Allows an app to toggle the incoming audio speaker setting for the meeting user from mute to unmute or vice-versa
-   * @param callback Callback contains 2 parameters, error and result.
    * error can either contain an error of type SdkError, incase of an error, or null when toggle is successful
    * result can either contain the true/false value, incase of a successful toggle or null when the toggling fails
-   * result: True means incoming audio is muted and false means incoming audio is unmuted
+   * @returns Promise result where true means incoming audio is muted and false means incoming audio is unmuted or rejected promise containing SdkError details
    */
-  export function toggleIncomingClientAudio(callback: (error: SdkError | null, result: boolean | null) => void): void {
-    if (!callback) {
-      throw new Error('[toggle incoming client audio] Callback cannot be null');
-    }
-    ensureInitialized(FrameContexts.sidePanel, FrameContexts.meetingStage);
-    sendMessageToParent('toggleIncomingClientAudio', callback);
+  export function toggleIncomingClientAudio(): Promise<boolean> {
+    return new Promise<boolean>(resolve => {
+      ensureInitialized(FrameContexts.sidePanel, FrameContexts.meetingStage);
+      resolve(sendAndHandleSdkError('toggleIncomingClientAudio'));
+    });
   }
 
   /**
    * @private
    * Hide from docs
    * Allows an app to get the meeting details for the meeting
-   * @param callback Callback contains 2 parameters, error and meetingDetails.
-   * error can either contain an error of type SdkError, incase of an error, or null when get is successful
-   * result can either contain a IMeetingDetails value, incase of a successful get or null when the get fails
+   * @returns Promise containing the meeting details in IMeetingDetails form or rejected promise containing SdkError details
    */
-  export function getMeetingDetails(
-    callback: (error: SdkError | null, meetingDetails: IMeetingDetails | null) => void,
-  ): void {
-    if (!callback) {
-      throw new Error('[get meeting details] Callback cannot be null');
-    }
-    ensureInitialized(
-      FrameContexts.sidePanel,
-      FrameContexts.meetingStage,
-      FrameContexts.settings,
-      FrameContexts.content,
-    );
-    sendMessageToParent('meeting.getMeetingDetails', callback);
+  export function getMeetingDetails(): Promise<IMeetingDetails> {
+    return new Promise<IMeetingDetails>(resolve => {
+      ensureInitialized(
+        FrameContexts.sidePanel,
+        FrameContexts.meetingStage,
+        FrameContexts.settings,
+        FrameContexts.content,
+      );
+      resolve(sendAndHandleSdkError('meeting.getMeetingDetails'));
+    });
   }
 
   /**
    * @private
    * Allows an app to get the authentication token for the anonymous or guest user in the meeting
-   * @param callback Callback contains 2 parameters, error and authenticationTokenOfAnonymousUser.
-   * error can either contain an error of type SdkError, incase of an error, or null when get is successful
-   * authenticationTokenOfAnonymousUser can either contain a string value, incase of a successful get or null when the get fails
+   * @returns Promise containing the token or rejected promise containing SdkError details
    */
-  export function getAuthenticationTokenForAnonymousUser(
-    callback: (error: SdkError | null, authenticationTokenOfAnonymousUser: string | null) => void,
-  ): void {
-    if (!callback) {
-      throw new Error('[get Authentication Token For AnonymousUser] Callback cannot be null');
-    }
-    ensureInitialized(FrameContexts.sidePanel, FrameContexts.meetingStage);
-    sendMessageToParent('meeting.getAuthenticationTokenForAnonymousUser', callback);
+  export function getAuthenticationTokenForAnonymousUser(): Promise<string> {
+    return new Promise<string>(resolve => {
+      ensureInitialized(FrameContexts.sidePanel, FrameContexts.meetingStage);
+      resolve(sendAndHandleSdkError('meeting.getAuthenticationTokenForAnonymousUser'));
+    });
   }
 
   export function isSupported(): boolean {
@@ -185,50 +175,39 @@ export namespace meeting {
 
   /**
    * Allows an app to get the state of the live stream in the current meeting
-   * @param callback Callback contains 2 parameters: error and liveStreamState.
-   * error can either contain an error of type SdkError, in case of an error, or null when get is successful
-   * liveStreamState can either contain a LiveStreamState value, or null when operation fails
+   * @returns Promise containing the LiveStreamState value or rejected promise containing SdkError details
    */
-  export function getLiveStreamState(
-    callback: (error: SdkError | null, liveStreamState: LiveStreamState | null) => void,
-  ): void {
-    if (!callback) {
-      throw new Error('[get live stream state] Callback cannot be null');
-    }
-    ensureInitialized(FrameContexts.sidePanel);
-    sendMessageToParent('meeting.getLiveStreamState', callback);
+  export function getLiveStreamState(): Promise<LiveStreamState> {
+    return new Promise<LiveStreamState>(resolve => {
+      ensureInitialized();
+      resolve(sendAndHandleSdkError('meeting.getLiveStreamState'));
+    });
   }
 
   /**
    * Allows an app to request the live streaming be started at the given streaming url
    * @param streamUrl the url to the stream resource
    * @param streamKey the key to the stream resource
-   * @param callback Callback contains error parameter which can be of type SdkError in case of an error, or null when operation is successful
    * Use getLiveStreamState or registerLiveStreamChangedHandler to get updates on the live stream state
+   * @returns Promise that will be resolved when the operation has completed or rejected with SdkError value
    */
-  export function requestStartLiveStreaming(
-    callback: (error: SdkError | null) => void,
-    streamUrl: string,
-    streamKey?: string,
-  ): void {
-    if (!callback) {
-      throw new Error('[request start live streaming] Callback cannot be null');
-    }
-    ensureInitialized(FrameContexts.sidePanel);
-    sendMessageToParent('meeting.requestStartLiveStreaming', [streamUrl, streamKey], callback);
+  export function requestStartLiveStreaming(streamUrl: string, streamKey?: string): Promise<void> {
+    return new Promise<void>(resolve => {
+      ensureInitialized(FrameContexts.sidePanel);
+      resolve(sendAndHandleSdkError('meeting.requestStartLiveStreaming', streamUrl, streamKey));
+    });
   }
 
   /**
    * Allows an app to request the live streaming be stopped at the given streaming url
-   * @param callback Callback contains error parameter which can be of type SdkError in case of an error, or null when operation is successful
    * Use getLiveStreamState or registerLiveStreamChangedHandler to get updates on the live stream state
+   * @returns Promise that will be resolved when the operation has completed or rejected with SdkError value
    */
-  export function requestStopLiveStreaming(callback: (error: SdkError | null) => void): void {
-    if (!callback) {
-      throw new Error('[request stop live streaming] Callback cannot be null');
-    }
-    ensureInitialized(FrameContexts.sidePanel);
-    sendMessageToParent('meeting.requestStopLiveStreaming', callback);
+  export function requestStopLiveStreaming(): Promise<void> {
+    return new Promise<void>(resolve => {
+      ensureInitialized(FrameContexts.sidePanel);
+      resolve(sendAndHandleSdkError('meeting.requestStopLiveStreaming'));
+    });
   }
 
   /**
@@ -247,18 +226,37 @@ export namespace meeting {
   /**
    * Allows an app to share contents in the meeting
    * @param callback Callback contains 2 parameters, error and result.
-   * error can either contain an error of type SdkError, incase of an error, or null when share is successful
-   * result can either contain a true value, incase of a successful share or null when the share fails
    * @param appContentUrl is the input URL which needs to be shared on to the stage
+   * @returns Promise resolved indicating whether or not the share was successful or rejected with SdkError value
    */
-  export function shareAppContentToStage(
-    callback: (error: SdkError | null, result: boolean | null) => void,
-    appContentUrl: string,
-  ): void {
-    if (!callback) {
-      throw new Error('[share app content to stage] Callback cannot be null');
-    }
-    ensureInitialized(FrameContexts.sidePanel);
-    sendMessageToParent('meeting.shareAppContentToStage', [appContentUrl], callback);
+  export function shareAppContentToStage(appContentUrl: string): Promise<boolean> {
+    return new Promise<boolean>(resolve => {
+      ensureInitialized(FrameContexts.sidePanel);
+      resolve(sendAndHandleSdkError('meeting.shareAppContentToStage', appContentUrl));
+    });
+  }
+
+  /**
+   * Provides information related app's in-meeting sharing capabilities
+   * @param callback Callback contains 2 parameters, error and result.
+   * @returns Promise resolved with sharing capability details or rejected with SdkError value
+   */
+  export function getAppContentStageSharingCapabilities(): Promise<IAppContentStageSharingCapabilities> {
+    return new Promise<IAppContentStageSharingCapabilities>(resolve => {
+      ensureInitialized(FrameContexts.sidePanel);
+      resolve(sendAndHandleSdkError('meeting.getAppContentStageSharingCapabilities'));
+    });
+  }
+
+  /**
+   * Terminates current stage sharing session in meeting
+   * @param callback Callback contains 2 parameters, error and result.
+   * @returns Promise resolved indicating whether or not sharing successfully stopped or rejected with SdkError value
+   */
+  export function stopSharingAppContentToStage(): Promise<boolean> {
+    return new Promise<boolean>(resolve => {
+      ensureInitialized(FrameContexts.sidePanel);
+      resolve(sendAndHandleSdkError('meeting.stopSharingAppContentToStage'));
+    });
   }
 }